Update sizeCC with author name and merge GraphGenerator
